requirements.txt
google-cloud-aiplatform[adk,agent_engines]
google-cloud-firestore
google-auth
requests
python-dotenv
scrapy
pydantic
google-cloud-bigquery
fastapi
uvicorn[standard]
google-generativeai
reportlab
google-cloud-storage